Week15/Hackaton/document_search/requirements.txt
sentence-transformers>=2.3.0
faiss-cpu>=1.8.0
transformers>=4.30.0
torch>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
streamlit>=1.28.0
rouge-score>=0.1.2
watchdog>=3.0.0
tqdm>=4.66.0
pycryptodome>=3.19.0
